requirements.txt
# API and Environment
python-dotenv==1.0.0
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-core==0.3.59
langgraph==0.4.3
pymongo>=4.6.0
langgraph-checkpoint-mongodb>=0.0.3

# Web Framework
flask==2.3.3
flask-socketio==5.3.6

# WSGI Server
gunicorn==21.2.0

# Utils
requests==2.31.0

# Diagram Generation
pyppeteer==2.0.0
grandalf==0.8 